Alerting3 (#19)
* This removes the alert tab in the edit from dashboards

Signed-off-by: Miek Gieben <miek@miek.nl>

* UI: remove the alerting tab

This removes the alerting tab from the UI in the edit dashboard. Still
to remove is the Alerting tab from the main UI and the Go code that does
stuff. Still unclear if the notifications are part of this nonsense or
are used to get notified of new dashboards (for instance);

Signed-off-by: Miek Gieben <miek@miek.nl>